Long

+++ BusinessConnectManagement/Areas/Faculty/Controllers/CooperationCategoriesController.cs
@@ -1,4 +1,5 @@
-﻿using BusinessConnectManagement.Models;
+﻿using BusinessConnectManagement.Middleware;
+using BusinessConnectManagement.Models;
 using System;
 using System.Collections.Generic;
 using System.Data.Entity;
@@ -9,7 +10,7 @@ using System.Web.Mvc;
 
 namespace BusinessConnectManagement.Areas.Faculty.Controllers
 {
-    [Authorize]
+    [LoginVerification]
     public class CooperationCategoriesController : Controller
     {
         private BCMEntities db = new BCMEntities();
@@ -17,7 +18,7 @@ namespace BusinessConnectManagement.Areas.Faculty.Controllers
         // GET: Faculty/CooperationCategories
         public ActionResult Index()
         {
-            return View(db.CooperationCategories.ToList());
+            return View(db.CooperationCategories.OrderByDescending(c => c.ID).ToList());
         }
 
         // GET: Faculty/CooperationCategories/Details/5
@@ -52,9 +53,13 @@ namespace BusinessConnectManagement.Areas.Faculty.Controllers
         {
             if (ModelState.IsValid)
             {
-                db.CooperationCategories.Add(cooperationCategorie);
-                db.SaveChanges();
-
+                if(db.CooperationCategories.Where(x => x.CooperationCategoriesName == cooperationCategorie.CooperationCategoriesName).Any() == true)
+                {
+                    TempData["AlertMessage"] = "<div class=\"toast toast--error\">\r\n     <div class=\"toast-left toast-left--error\">\r\n       <i class=\"fas fa-times-circle\"></i>\r\n     </div>\r\n     <div class=\"toast-content\">\r\n       <p class=\"toast-text\">Tên danh mục hợp tác đã tồn tại.</p>\r\n     </div>\r\n     <div class=\"toast-right\">\r\n       <i style=\"cursor:pointer\" class=\"toast-icon fas fa-times\" onclick=\"remove()\"></i>\r\n     </div>\r\n   </div>";
+                    return RedirectToAction("Index");
+                }
+                db.CooperationCategories.Add(cooperationCategorie); db.SaveChanges();
+                TempData["AlertMessage"] = "<div class=\"toast toast--success\">\r\n     <div class=\"toast-left toast-left--success\">\r\n       <i class=\"fas fa-check-circle\"></i>\r\n     </div>\r\n     <div class=\"toast-content\">\r\n       <p class=\"toast-text\">Thêm dữ liệu thành công.</p>\r\n     </div>\r\n     <div class=\"toast-right\">\r\n      <i style=\"cursor:pointer\" class=\"toast-icon fas fa-times\" onclick=\"remove()\"></i>\r\n     </div>\r\n   </div>\r\n";
                 return RedirectToAction("Index");
             }
 
@@ -89,7 +94,7 @@ namespace BusinessConnectManagement.Areas.Faculty.Controllers
             {
                 db.Entry(cooperationCategorie).State = EntityState.Modified;
                 db.SaveChanges();
-
+                TempData["AlertMessage"] = "<div class=\"toast toast--success\">\r\n     <div class=\"toast-left toast-left--success\">\r\n       <i class=\"fas fa-check-circle\"></i>\r\n     </div>\r\n     <div class=\"toast-content\">\r\n       <p class=\"toast-text\">Cập nhật thành công.</p>\r\n     </div>\r\n     <div class=\"toast-right\">\r\n      <i style=\"cursor:pointer\" class=\"toast-icon fas fa-times\" onclick=\"remove()\"></i>\r\n     </div>\r\n   </div>\r\n";
                 return RedirectToAction("Index");
             }
 
@@ -121,9 +126,16 @@ namespace BusinessConnectManagement.Areas.Faculty.Controllers
         {
             CooperationCategory cooperationCategorie = db.CooperationCategories.Find(id);
 
+            if (db.BusinessCooperationCategories.Where(x => x.CooperationCategories_ID == id).Any() == true)
+            {
+                TempData["AlertMessage"] = "<div class=\"toast toast--error\">\r\n     <div class=\"toast-left toast-left--error\">\r\n       <i class=\"fas fa-times-circle\"></i>\r\n     </div>\r\n     <div class=\"toast-content\">\r\n       <p class=\"toast-text\">Xóa không thành công vì đã có doanh nghiệp chọn hình thưc hợp tác này.</p>\r\n     </div>\r\n     <div class=\"toast-right\">\r\n       <i style=\"cursor:pointer\" class=\"toast-icon fas fa-times\" onclick=\"remove()\"></i>\r\n     </div>\r\n   </div>";
+
+                return RedirectToAction("Index");
+            }
+
             db.CooperationCategories.Remove(cooperationCategorie);
             db.SaveChanges();
-
+            TempData["AlertMessage"] = "<div class=\"toast toast--success\">\r\n     <div class=\"toast-left toast-left--success\">\r\n       <i class=\"fas fa-check-circle\"></i>\r\n     </div>\r\n     <div class=\"toast-content\">\r\n       <p class=\"toast-text\">Xóa thành công.</p>\r\n     </div>\r\n     <div class=\"toast-right\">\r\n      <i style=\"cursor:pointer\" class=\"toast-icon fas fa-times\" onclick=\"remove()\"></i>\r\n     </div>\r\n   </div>\r\n";
             return RedirectToAction("Index");
         }
